requirements.txt
beautifulsoup4==4.14.3
certifi==2026.2.25
charset-normalizer==3.4.7
chromadb==1.5.8
click==8.3.3
google-genai==1.73.1
idna==3.11
numpy==2.2.4
pymupdf==1.25.3
rank-bm25==0.2.2
sentence-transformers==3.4.1
requests==2.33.1
soupsieve==2.8.3
typing_extensions==4.15.0
urllib3==2.6.3
